# Avoid allocation in JSON.Scan with string