`[]byte(src.(string))` is an unneeded allocation, because you can splat a string directly into an append call.

Benchmark:

```
$ go test -benchmem -bench=JSON

# Before

goos: darwin
goarch: arm64
pkg: github.com/volatiletech/sqlboiler/v4/types
BenchmarkJSON_Scan-8   	 4658832	       237.9 ns/op	    2304 B/op	       2 allocs/op
PASS
ok  	github.com/volatiletech/sqlboiler/v4/types	1.738s

# After

goos: darwin
goarch: arm64
pkg: github.com/volatiletech/sqlboiler/v4/types
BenchmarkJSON_Scan-8   	 8657862	       118.0 ns/op	    1152 B/op	       1 allocs/op
PASS
ok  	github.com/volatiletech/sqlboiler/v4/types	1.435s
```