@aarondl 2021-01-24
Ah, great change! But the target branch should be dev as per CONTRIBUTING.md. Sorry to get to this so late :)